[review] Moved read::pe::PeFile::export_table to another file
