+++ src/read/pe/file.rs
@@ -9,9 +9,7 @@ use crate::read::{
     self, Architecture, ComdatKind, Error, Export, FileFlags, Import, NoDynamicRelocationIterator,
     Object, ObjectComdat, ReadError, ReadRef, Result, SectionIndex, SymbolIndex,
 };
-use crate::{
-    pe, ByteString, Bytes, CodeView, LittleEndian as LE, Pod, U16Bytes, U32Bytes, U32, U64,
-};
+use crate::{pe, ByteString, Bytes, CodeView, LittleEndian as LE, Pod, U32, U64};
 
 use super::Export as PeExport;
 use super::{PeSection, PeSectionIterator, PeSegment, PeSegmentIterator, SectionTable};
@@ -95,120 +93,6 @@ where
     /// Returns this binary data.
     pub fn data(&self) -> R {
         self.data
-    }
-
-    /// Returns the exports of this PE file
-    ///
-    /// See also the [`PeFile::exports`] function, which only returns a subset of these exports.
-    pub fn export_table(&self) -> Result<Vec<PeExport<'data>>> {
-        let data_dir = match self.data_directory(pe::IMAGE_DIRECTORY_ENTRY_EXPORT) {
-            Some(data_dir) => data_dir,
-            None => return Ok(Vec::new()),
-        };
-        let export_va = data_dir.virtual_address.get(LE);
-        let export_size = data_dir.size.get(LE);
-        let export_data = data_dir.data(self.data, &self.common.sections).map(Bytes)?;
-        let export_dir = export_data
-            .read_at::<pe::ImageExportDirectory>(0)
-            .read_error("Invalid PE export dir size")?;
-        let addresses = export_data
-            .read_slice_at::<U32Bytes<_>>(
-                export_dir
-                    .address_of_functions
-                    .get(LE)
-                    .wrapping_sub(export_va) as usize,
-                export_dir.number_of_functions.get(LE) as usize,
-            )
-            .read_error("Invalid PE export address table")?;
-        let number = export_dir.number_of_names.get(LE) as usize;
-        let names = export_data
-            .read_slice_at::<U32Bytes<_>>(
-                export_dir.address_of_names.get(LE).wrapping_sub(export_va) as usize,
-                number,
-            )
-            .read_error("Invalid PE export name table")?;
-        let base_ordinal = export_dir.base.get(LE);
-        let ordinals = export_data
-            .read_slice_at::<U16Bytes<_>>(
-                export_dir
-                    .address_of_name_ordinals
-                    .get(LE)
-                    .wrapping_sub(export_va) as usize,
-                number,
-            )
-            .read_error("Invalid PE export ordinal table")?;
-
-        // First, let's list all exports...
-        let mut exports = Vec::new();
-        for (i, address) in addresses.iter().enumerate() {
-            // Convert from an array index to an ordinal
-            // The MSDN documentation is wrong here, see https://stackoverflow.com/a/40001778/721832
-            let ordinal: u32 = match i.try_into() {
-                Err(_err) => continue,
-                Ok(index) => index,
-            };
-            let ordinal = ordinal + base_ordinal;
-            let address = address.get(LE);
-
-            // is it a regular or forwarded export?
-            if address < export_va || (address - export_va) >= export_size {
-                exports.push(PeExport::ByOrdinal {
-                    ordinal: ordinal,
-                    address: self.common.image_base.wrapping_add(address as u64),
-                });
-            } else {
-                let forwarded_to = export_data
-                    .read_string_at(address.wrapping_sub(export_va) as usize)
-                    .read_error("Invalid target for PE forwarded export")?;
-                exports.push(PeExport::ForwardedByOrdinal {
-                    ordinal: ordinal,
-                    forwarded_to: forwarded_to,
-                });
-            }
-        }
-
-        // Now, check whether some (or all) of them have an associated name
-        for (name_ptr, ordinal_index) in names.iter().zip(ordinals.iter()) {
-            // Items in the ordinal array are biased.
-            // The MSDN documentation is wrong regarding this bias, see https://stackoverflow.com/a/40001778/721832
-            let ordinal_index = ordinal_index.get(LE) as u32;
-
-            let name = export_data
-                .read_string_at(name_ptr.get(LE).wrapping_sub(export_va) as usize)
-                .read_error("Invalid PE export name entry")?;
-
-            let unnamed_equivalent = exports.get(ordinal_index as usize).cloned();
-            match unnamed_equivalent {
-                Some(PeExport::ByOrdinal { ordinal, address }) => {
-                    let _ = core::mem::replace(
-                        &mut exports[ordinal_index as usize],
-                        PeExport::Regular {
-                            name,
-                            address,
-                            ordinal,
-                        },
-                    );
-                }
-
-                Some(PeExport::ForwardedByOrdinal {
-                    ordinal,
-                    forwarded_to,
-                }) => {
-                    let _ = core::mem::replace(
-                        &mut exports[ordinal_index as usize],
-                        PeExport::Forwarded {
-                            name,
-                            ordinal,
-                            forwarded_to,
-                        },
-                    );
-                }
-
-                _ => continue, // unless ordinals are not unique in the ordinals array, this should not happen
-            }
-        }
-
-        Ok(exports)
     }
 }
